Implemented a MongDB version of the app with unit tests.
This is pretty bad because the unit tests are connecting to an actual MongoDB instance in the cloud (Atlas). There's some in-memory Spring thing that's theoretically possible but I couldn't get it to work and I wanted to get the logic working. This actually won't build on docker this way (I think) and it's just bad practice anyway (and it's really slow btw). Supposedly running a local instance is the "best" way to do this if you really want to emulate prod and I've done that before in past projects but I want something that will work on my Linux VM which is where I need to compile this to run it on K8s. So next step is to get these tests working on a local MongoDB, preferably within a docker container.

I only unit tested this btw. I don't know if this will actually work from the API yet.