Change all URL paths so that /vote is the base URL.
It's like this for prod ingress configs. The options I was considering was either using a new subdomain (i.e., vote.agaber.dev) or a separate root URL agaber.dev/vote/* because I want agaber.dev for other stuff. I opted for the latter option but it required changing a bunch of stuff in the UI code.